mixer_paths: Add controls to forbid sending afe cal for proxy ports
Update mixer paths for all variants of msm8952 and msm8976
with 2 new mixer controls to provide a mechanism to permit
or prevent sending of afe cal for proxy ports.

Currently there is no valid afe cal for proxy ports, so forbid
sending of afe cal to DSP for proxy ports to prevent
port start failures.

CRs-Fixed: 880657
Change-Id: I2cf5d344a2288be5ddd872d2462f1eba5c442319